Extract shared directives into a shared base class
This commit extracts directives that are shared between multiple object
types into a shared_language module.
