+++ lib/ramble/ramble/appkit.py
@@ -23,5 +23,6 @@ from ramble.spec import Spec
 
 import ramble.language.application_language
 from ramble.language.application_language import *
+from ramble.language.shared_language import *
 
 from ramble.schema.types import OUTPUT_CAPTURE
