+++ lib/ramble/ramble/modkit.py
@@ -21,4 +21,5 @@ from ramble.spec import Spec
 
 import ramble.language.modifier_language
 from ramble.language.modifier_language import *
+from ramble.language.shared_language import *
 from ramble.schema.types import OUTPUT_CAPTURE
